refacto: expe.py splited
Changelog:
- Due to a circular dependencies I choose take the opportunity to split expe.py.
  I moved all type declaration into base/data_type.py

- Now the pipeline have default input/output directory flow that can be overwritten by the user
