+++ src/ragtime/base/retriever.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 
 from abc import ( abstractmethod )
-from ragtime.expe import ( RagtimeBase, QA )
+from ragtime.base.data_type import (
+    RagtimeBase,
+    QA
+)
 
 class Retriever(RagtimeBase):
     """
